Use one flow for data and configuration. Fix hanging UI on fast/many streams.
I decided to use one flow for configuration and data because the resulting code
is much easier to read and reason about. As long as the order of
1. start inlet -> 2. start collecting is kept, there should be no loss of packages.

To prevent the UI from hanging if there are a lot of samples coming in,
I moved the handling of data events out of the UI thread. As the chart
data is kept as StateFlow, this should not introduce any synchronisation
problems.